requirements.txt
# Campaign Respond Dependencies
anthropic>=0.40.0
python-docx>=1.1.0
PyPDF2>=3.0.0
markdown>=3.5.0
weasyprint>=62.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
msal>=1.28.0
python-dotenv>=1.0.0
jinja2>=3.1.0
httpx>=0.27.0
